spring/src/main/java/com/spittr/model/Spittle.java
package com.spittr.model;

import java.util.Date;

public class Spittle {
    private final Long id;
    private final String message;
    private final Date time;
    private Double latitude;
    private Double longitude;

    public Spittle(long id, String message, Date time) {
        this(id, message, time, null, null);
    }

    public Spittle(
            long id, String message, Date time, Double longitude, Double latitude) {
        this.id = id;
        this.message = message;
        this.time = time;
        this.longitude = longitude;
        this.latitude = latitude;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;

        if (o == null || getClass() != o.getClass()) return false;

        Spittle spittle = (Spittle) o;

        return new org.apache.commons.lang.builder.EqualsBuilder()
                .append(time, spittle.time)
                .append(message, spittle.message)
                .append(latitude, spittle.latitude)
                .append(longitude, spittle.longitude)
                .isEquals();
    }

    @Override
    public int hashCode() {
        return new org.apache.commons.lang.builder.HashCodeBuilder(17, 37)
                .append(time)
                .append(message)
                .append(latitude)
                .append(longitude)
                .toHashCode();
    }

    public Long getId() {
        return id;
    }

    public String getMessage() {
        return message;
    }

    public Date getTime() {
        return time;
    }

    public Double getLatitude() {
        return latitude;
    }

    public void setLatitude(Double latitude) {
        this.latitude = latitude;
    }

    public Double getLongitude() {
        return longitude;
    }

    public void setLongitude(Double longitude) {
        this.longitude = longitude;
    }
}
